+++ web/src/config/chainConfigs/ChainConfigsDusk.ts
@@ -1,4 +1,4 @@
-import type { EvmChainInfo, EvmChains, IbcChainInfo, IbcChains } from ".";
+import type { EvmChainInfo, EvmChains, IbcChainInfo, IbcChains } from "./types";
 
 const CelestiaChainInfo: IbcChainInfo = {
   // Chain-id of the celestia chain.
@@ -192,17 +192,19 @@ const FlameChainInfo: EvmChainInfo = {
       coinDenom: "RIA",
       coinMinimalDenom: "uria",
       coinDecimals: 18,
+      ibcWithdrawalFeeWei: "10000000000000000",
       iconClass: "i-celestia",
     },
     {
       coinDenom: "USDC",
       coinMinimalDenom: "uusdc",
       coinDecimals: 18,
-      iconClass: "i-noble",
       // address of erc20 contract on dusk-11
       erc20ContractAddress: "0xa4f59B3E97EC22a2b949cB5b6E8Cd6135437E857",
       // this value would only exist for native tokens
       nativeTokenWithdrawerContractAddress: "",
+      ibcWithdrawalFeeWei: "10000000000000000",
+      iconClass: "i-noble",
     },
     {
       coinDenom: "fakeTIA",
@@ -212,6 +214,7 @@ const FlameChainInfo: EvmChainInfo = {
       //  just using this for testing the UI.
       erc20ContractAddress: "0xFc83F6A786728F448481B7D7d5C0659A92a62C4d",
       nativeTokenWithdrawerContractAddress: "",
+      ibcWithdrawalFeeWei: "10000000000000000",
       iconClass: "i-celestia",
     },
   ],
